# docs/swagger.yaml
basePath: /api/v1
definitions:
  dao.User:
    properties:
      ccnuid:
        type: string
      id:
        type: string
      name:
        type: string
      passwd:
        type: string
      role:
        description: 'role: 0 for student, 1 for teacher, 2 for school'
        type: integer
      school:
        type: string
      stage:
        type: string
      uid:
        type: string
      valid:
        type: integer
    type: object
  handler.CheckFormRequest:
    properties:
      advice:
        type: string
      fid:
        type: string
      opt:
        type: integer
    type: object
  handler.FormRequest:
    properties:
      college:
        type: string
      community_advice:
        type: string
      contact:
        type: string
      context:
        type: string
      create_at:
        type: string
      from_bed:
        type: string
      from_dorm:
        type: string
      hqzb_advice:
        type: string
      id:
        type: string
      status:
        type: integer
      student_id:
        type: string
      teacher_advice:
        type: string
      teacher_id:
        type: string
      to_bed:
        type: string
      to_dorm:
        type: string
      xgb_advice:
        type: string
    type: object
  handler.LoginRequest:
    properties:
      ccnuid:
        type: string
      name:
        type: string
      passwd:
        type: string
    type: object
  handler.PDFRequest:
    properties:
      contact:
        type: string
      from:
        type: string
      name:
        type: string
      school:
        type: string
      sid:
        type: string
      text0:
        type: string
      text1:
        type: string
      text2:
        type: string
      text3:
        type: string
      text4:
        type: string
      to:
        type: string
    type: object
  handler.RegisterRequest:
    properties:
      ccnuid:
        type: string
      name:
        type: string
      passwd:
        type: string
    type: object
  handler.ReportRequest:
    properties:
      contact:
        type: string
      info:
        type: string
    type: object
  handler.Resp:
    properties:
      code:
        type: integer
      data: {}
      msg: {}
    type: object
info:
  contact:
    email: 3557695455@qq.com
    name: '@a48zhang'
  description: Backend system of Dorm-changing mini program, CCNU
  title: Dorm-changing Backend API
  version: "0.3"
paths:
  /forms:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: fid
        in: query
        name: fid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 获取单个换宿申请表格
      tags:
      - form
  /forms/assigned:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 获取分配给指定角色的换宿申请表格
      tags:
      - form
  /forms/check:
    post:
      consumes:
      - application/json
      description: |-
        opt: 0 for waiting, 1 for teacher approve, -1 for teacher reject, 2 for school approve, -2 for school reject
        if role < abs(opt), permission denied
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 审核信息
        in: body
        name: CheckForm
        required: true
        schema:
          $ref: '#/definitions/handler.CheckFormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Resp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 审核换宿申请表格
      tags:
      - form
  /forms/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 换宿申请表格
        in: body
        name: Form
        required: true
        schema:
          $ref: '#/definitions/handler.FormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 提交新换宿申请表格
      tags:
      - form
  /forms/my:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 获取用户的换宿申请表格
      tags:
      - form
  /forms/pdf:
    get:
      consumes:
      - application/json
      description: Generate PDF for form by ID
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: form id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 根据表单ID生成表单PDF
      tags:
      - form
    post:
      consumes:
      - application/json
      description: Generate PDF for form
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/handler.PDFRequest'
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 根据上传数据生成表单PDF
      tags:
      - form
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息，只需要填写Name和Passwd
        in: body
        name: Login
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 登录
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息
        in: body
        name: Register
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 注册
      tags:
      - auth
  /report:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 我的反馈
      tags:
      - user
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 反馈信息
        in: body
        name: Report
        required: true
        schema:
          $ref: '#/definitions/handler.ReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 提交新反馈
      tags:
      - user
  /users:
    get:
      consumes:
      - application/json
      description: Get user info
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 获取用户信息
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Update user's info
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/dao.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 修改用户信息
      tags:
      - user
schemes:
- http
swagger: "2.0"